Move bulk ES sync code to separate module
DH-1660

As it's no longer specific to the management command.
